+++ src/code42cli/cmds/devices.py
@@ -1,16 +1,19 @@
 from datetime import date
-from datetime import datetime
 
 import click
 from pandas import concat
 from pandas import DataFrame
 from pandas import to_datetime
-from pandas import to_timedelta
 from py42 import exceptions
 from py42.exceptions import Py42NotFoundError
 
+from code42cli.bulk import generate_template_cmd_factory
 from code42cli.bulk import run_bulk_process
 from code42cli.click_ext.groups import OrderedGroup
+from code42cli.click_ext.options import incompatible_with
+from code42cli.click_ext.types import MagicDate
+from code42cli.date_helper import round_datetime_to_day_end
+from code42cli.date_helper import round_datetime_to_day_start
 from code42cli.errors import Code42CLIError
 from code42cli.file_readers import read_csv_arg
 from code42cli.options import format_option
@@ -26,16 +29,20 @@ def devices(state):
     pass
 
 
-device_guid_argument = click.argument("device-guid", type=str)
-
-change_device_name_option = click.option(
-    "--change-device-name",
-    required=False,
-    is_flag=True,
-    default=False,
-    help="""Prepend "deactivated_" and today's date to the name of any
-    deactivated devices.""",
+device_guid_argument = click.argument(
+    "device-guid", type=str, callback=lambda ctx, param, arg: _verify_guid_type(arg),
 )
+
+
+def change_device_name_option(help_msg):
+    return click.option(
+        "--change-device-name",
+        required=False,
+        is_flag=True,
+        default=False,
+        help=help_msg,
+    )
+
 
 DATE_FORMAT = "%Y-%m-%d"
 purge_date_option = click.option(
@@ -43,14 +50,16 @@ purge_date_option = click.option(
     required=False,
     type=click.DateTime(formats=[DATE_FORMAT]),
     default=None,
-    help="""The date on which the archive should be purged from cold storage in yyyy-MM-dd format.
-    If not provided, the date will be set according to the appropriate org settings.""",
+    help="The date on which the archive should be purged from cold storage in yyyy-MM-dd format. "
+    "If not provided, the date will be set according to the appropriate org settings.",
 )
 
 
 @devices.command()
 @device_guid_argument
-@change_device_name_option
+@change_device_name_option(
+    "Prepend 'deactivated_<current_date>' to the name of the device if deactivation is successful."
+)
 @purge_date_option
 @sdk_options()
 def deactivate(state, device_guid, change_device_name, purge_date):
@@ -58,16 +67,21 @@ def deactivate(state, device_guid, change_device_name, purge_date):
     _deactivate_device(state.sdk, device_guid, change_device_name, purge_date)
 
 
+@devices.command()
+@device_guid_argument
+@sdk_options()
+def reactivate(state, device_guid):
+    """Reactivate a device within Code42. Requires the device GUID to reactivate."""
+    _reactivate_device(state.sdk, device_guid)
+
+
 def _deactivate_device(sdk, device_guid, change_device_name, purge_date):
-    device = sdk.devices.get_by_guid(device_guid)
     try:
-        sdk.devices.deactivate(device.data["computerId"])
+        device = _change_device_activation(sdk, device_guid, "deactivate")
     except exceptions.Py42BadRequestError:
-        raise Code42CLIError("The device {} is in legal hold.".format(device_guid))
-    except exceptions.Py42NotFoundError:
-        raise Code42CLIError("The device {} was not found.".format(device_guid))
-    except exceptions.Py42ForbiddenError:
-        raise Code42CLIError("Unable to deactivate {}.".format(device_guid))
+        raise Code42CLIError(
+            "The device with GUID '{}' is in legal hold.".format(device_guid)
+        )
     if purge_date:
         _update_cold_storage_purge_date(sdk, device_guid, purge_date)
     if change_device_name and not device.data["name"].startswith("deactivated_"):
@@ -79,6 +93,39 @@ def _deactivate_device(sdk, device_guid, change_device_name, purge_date):
             + "_"
             + device.data["name"],
         )
+
+
+def _reactivate_device(sdk, device_guid):
+    _change_device_activation(sdk, device_guid, "reactivate")
+
+
+def _change_device_activation(sdk, device_guid, cmd_str):
+    try:
+        device = sdk.devices.get_by_guid(device_guid)
+        device_id = device.data["computerId"]
+        if cmd_str == "reactivate":
+            sdk.devices.reactivate(device_id)
+        elif cmd_str == "deactivate":
+            sdk.devices.deactivate(device_id)
+        return device
+    except exceptions.Py42NotFoundError:
+        raise Code42CLIError(
+            "The device with GUID '{}' was not found.".format(device_guid)
+        )
+    except exceptions.Py42ForbiddenError:
+        raise Code42CLIError(
+            "Unable to {} the device with GUID '{}'.".format(cmd_str, device_guid)
+        )
+
+
+def _verify_guid_type(device_guid):
+    if device_guid is None:
+        return
+    try:
+        int(device_guid)
+        return device_guid
+    except ValueError:
+        raise Code42CLIError("Not a valid guid.")
 
 
 def _update_cold_storage_purge_date(sdk, guid, purge_date):
@@ -103,17 +150,18 @@ def _change_device_name(sdk, guid, name):
 
 @devices.command()
 @device_guid_argument
-@format_option
 @sdk_options()
 def show(state, device_guid, format=None):
-    """Print device info. Requires device GUID."""
+    """Print individual device info. Requires device GUID."""
 
     formatter = OutputFormatter(format, _device_info_keys_map())
     backup_set_formatter = OutputFormatter(format, _backup_set_keys_map())
     device_info = _get_device_info(state.sdk, device_guid)
     formatter.echo_formatted_list([device_info])
-    click.echo()
-    backup_set_formatter.echo_formatted_list(device_info["backupUsage"])
+    backup_usage = device_info.get("backupUsage")
+    if backup_usage:
+        click.echo()
+        backup_set_formatter.echo_formatted_list(backup_usage)
 
 
 def _device_info_keys_map():
@@ -145,20 +193,23 @@ def _get_device_info(sdk, device_guid):
 
 active_option = click.option(
     "--active",
-    required=False,
-    type=bool,
     is_flag=True,
+    help="Limits results to only active devices.",
     default=None,
-    help="Get only active or deactivated devices. Defaults to getting all devices.",
 )
-
+inactive_option = click.option(
+    "--inactive",
+    is_flag=True,
+    help="Limits results to only deactivated devices.",
+    cls=incompatible_with("active"),
+)
 org_uid_option = click.option(
     "--org-uid",
     required=False,
     type=str,
     default=None,
-    help="""Limit devices to only the ones in the org you specify.
-    Note that child orgs will be included.""",
+    help="Limit devices to only the ones in the org you specify. "
+    "Note that child orgs will be included.",
 )
 
 include_usernames_option = click.option(
@@ -171,31 +222,18 @@ include_usernames_option = click.option(
 )
 
 
-@devices.command(name="list", help="Get information about many devices")
+@devices.command(name="list")
 @active_option
-@click.option(
-    "--days-since-last-connected",
-    required=False,
-    type=int,
-    help="Return only devices that have not connected in the number of days specified.",
-)
+@inactive_option
 @org_uid_option
-@click.option(
-    "--drop-most-recent",
-    required=False,
-    type=int,
-    help="""Will drop the X most recently connected devices for each user from the
-    result list where X is the number you provide as this argument. Can be used to
-    avoid passing the most recently connected device for a user to the deactivate command.""",
-)
 @click.option(
     "--include-backup-usage",
     required=False,
     type=bool,
     default=False,
     is_flag=True,
-    help="""Return backup usage information for each device
-    (may significantly lengthen the size of the return).""",
+    help="Return backup usage information for each device (may significantly lengthen the size "
+    "of the return).",
 )
 @include_usernames_option
 @click.option(
@@ -204,22 +242,59 @@ include_usernames_option = click.option(
     type=bool,
     default=False,
     is_flag=True,
-    help="""Include device settings in output.""",
+    help="Include device settings in output.",
+)
+@click.option(
+    "--exclude-most-recently-connected",
+    type=int,
+    help="Filter out the N most recently connected devices per user. "
+    "Useful for identifying duplicate and/or replaced devices that are no longer needed across "
+    "an environment. If a user has 2 devices and N=1, the one device with the most recent "
+    "'lastConnected' date will not show up in the result list.",
+)
+@click.option(
+    "--last-connected-before",
+    type=MagicDate(rounding_func=round_datetime_to_day_start),
+    help=f"Include devices only when the 'lastConnected' field is after the provided value. {MagicDate.HELP_TEXT}",
+)
+@click.option(
+    "--last-connected-after",
+    type=MagicDate(rounding_func=round_datetime_to_day_end),
+    help="Include devices only when 'lastConnected' field is after the provided value. "
+    "Argument format options are the same as --last-connected-before.",
+)
+@click.option(
+    "--created-before",
+    type=MagicDate(rounding_func=round_datetime_to_day_start),
+    help="Include devices only when 'creationDate' field is less than the provided value. "
+    "Argument format options are the same as --last-connected-before.",
+)
+@click.option(
+    "--created-after",
+    type=MagicDate(rounding_func=round_datetime_to_day_end),
+    help="Include devices only when 'creationDate' field is greater than the provided value. "
+    "Argument format options are the same as --last-connected-before.",
 )
 @format_option
 @sdk_options()
 def list_devices(
     state,
     active,
-    days_since_last_connected,
-    drop_most_recent,
+    inactive,
     org_uid,
     include_backup_usage,
     include_usernames,
     include_settings,
+    exclude_most_recently_connected,
+    last_connected_after,
+    last_connected_before,
+    created_after,
+    created_before,
     format,
 ):
-    """Outputs a list of all devices."""
+    """Get information about many devices."""
+    if inactive:
+        active = False
     columns = [
         "computerId",
         "guid",
@@ -227,30 +302,39 @@ def list_devices(
         "osHostname",
         "status",
         "lastConnected",
+        "creationDate",
         "productVersion",
         "osName",
         "osVersion",
         "userUid",
     ]
-    devices_dataframe = _get_device_dataframe(
+    df = _get_device_dataframe(
         state.sdk, columns, active, org_uid, include_backup_usage
     )
-    if drop_most_recent:
-        devices_dataframe = _drop_n_devices_per_user(
-            devices_dataframe, drop_most_recent
+    if last_connected_after:
+        df = df.loc[to_datetime(df.lastConnected) > last_connected_after]
+    if last_connected_before:
+        df = df.loc[to_datetime(df.lastConnected) < last_connected_before]
+    if created_after:
+        df = df.loc[to_datetime(df.creationDate) > created_after]
+    if created_before:
+        df = df.loc[to_datetime(df.creationDate) < created_before]
+    if exclude_most_recently_connected:
+        most_recent = (
+            df.sort_values(["userUid", "lastConnected"], ascending=False)
+            .groupby("userUid")
+            .head(exclude_most_recently_connected)
         )
-    if days_since_last_connected:
-        devices_dataframe = _drop_devices_which_have_not_connected_in_some_number_of_days(
-            devices_dataframe, days_since_last_connected
-        )
+        df = df.drop(most_recent.index)
     if include_settings:
-        devices_dataframe = _add_settings_to_dataframe(state.sdk, devices_dataframe)
+        df = _add_settings_to_dataframe(state.sdk, df)
     if include_usernames:
-        devices_dataframe = _add_usernames_to_device_dataframe(
-            state.sdk, devices_dataframe
-        )
-    formatter = DataFrameOutputFormatter(format)
-    formatter.echo_formatted_dataframe(devices_dataframe)
+        df = _add_usernames_to_device_dataframe(state.sdk, df)
+    if df.empty:
+        click.echo("No results found.")
+    else:
+        formatter = DataFrameOutputFormatter(format)
+        formatter.echo_formatted_dataframe(df)
 
 
 def _get_device_dataframe(
@@ -297,35 +381,6 @@ def _add_settings_to_dataframe(sdk, device_dataframe):
         return device_dataframe
 
 
-def _drop_devices_which_have_not_connected_in_some_number_of_days(
-    devices_dataframe, days_since_last_connected
-):
-    utc_now = to_datetime(datetime.utcnow(), utc=True)
-    devices_last_connected_dates = to_datetime(
-        devices_dataframe["lastConnected"], utc=True
-    )
-    days_since_last_connected_delta = to_timedelta(
-        days_since_last_connected, unit="days"
-    )
-    return devices_dataframe.loc[
-        utc_now - devices_last_connected_dates > days_since_last_connected_delta, :,
-    ]
-
-
-def _drop_n_devices_per_user(
-    device_dataframe,
-    number_to_drop,
-    sort_field="lastConnected",
-    sort_ascending=False,
-    group_field="userUid",
-):
-    return (
-        device_dataframe.sort_values(by=sort_field, ascending=sort_ascending)
-        .drop(device_dataframe.groupby(group_field).head(number_to_drop).index)
-        .reset_index(drop=True)
-    )
-
-
 def _add_usernames_to_device_dataframe(sdk, device_dataframe):
     users_generator = sdk.users.get_all()
     users_list = []
@@ -337,30 +392,29 @@ def _add_usernames_to_device_dataframe(sdk, device_dataframe):
     return device_dataframe.merge(users_dataframe, how="left", on="userUid")
 
 
-@devices.command(
-    name="list-backup-sets",
-    help="Get information about many devices and their backup sets",
-)
+@devices.command()
 @active_option
+@inactive_option
 @org_uid_option
 @include_usernames_option
 @format_option
 @sdk_options()
 def list_backup_sets(
-    state, active, org_uid, include_usernames, format,
+    state, active, inactive, org_uid, include_usernames, format,
 ):
-    """Outputs a list of all devices."""
+    """Get information about many devices and their backup sets."""
+    if inactive:
+        active = False
     columns = ["guid", "userUid"]
-    devices_dataframe = _get_device_dataframe(state.sdk, columns, active, org_uid)
+    df = _get_device_dataframe(state.sdk, columns, active, org_uid)
     if include_usernames:
-        devices_dataframe = _add_usernames_to_device_dataframe(
-            state.sdk, devices_dataframe
-        )
-    devices_dataframe = _add_backup_set_settings_to_dataframe(
-        state.sdk, devices_dataframe
-    )
-    formatter = DataFrameOutputFormatter(format)
-    formatter.echo_formatted_dataframe(devices_dataframe)
+        df = _add_usernames_to_device_dataframe(state.sdk, df)
+    df = _add_backup_set_settings_to_dataframe(state.sdk, df)
+    if df.empty:
+        click.echo("No results found.")
+    else:
+        formatter = DataFrameOutputFormatter(format)
+        formatter.echo_formatted_dataframe(df)
 
 
 def _add_backup_set_settings_to_dataframe(sdk, devices_dataframe):
@@ -414,17 +468,30 @@ def bulk(state):
     pass
 
 
-@bulk.command(
-    name="deactivate",
-    help="""Deactivate all devices on the given list.
-            Takes as input a CSV with a 'guid' column.""",
+_bulk_device_activation_headers = ["guid"]
+
+
+devices_generate_template = generate_template_cmd_factory(
+    group_name="devices",
+    commands_dict={
+        "reactivate": _bulk_device_activation_headers,
+        "deactivate": _bulk_device_activation_headers,
+    },
+    help_message="Generate the CSV template needed for bulk device commands.",
 )
-@read_csv_arg(headers=["guid"])
-@change_device_name_option
+bulk.add_command(devices_generate_template)
+
+
+@bulk.command(name="deactivate")
+@read_csv_arg(headers=_bulk_device_activation_headers)
+@change_device_name_option(
+    "Prepend 'deactivated_<current_date>' to the name of any successfully deactivated devices."
+)
 @purge_date_option
 @format_option
 @sdk_options()
 def bulk_deactivate(state, csv_rows, change_device_name, purge_date, format):
+    """Deactivate all devices from the provided CSV containing a 'guid' column."""
     sdk = state.sdk
     csv_rows[0]["deactivated"] = False
     formatter = OutputFormatter(format, {key: key for key in csv_rows[0].keys()})
@@ -444,5 +511,29 @@ def bulk_deactivate(state, csv_rows, change_device_name, purge_date, format):
 
     result_rows = run_bulk_process(
         handle_row, csv_rows, progress_label="Deactivating devices:"
+    )
+    formatter.echo_formatted_list(result_rows)
+
+
+@bulk.command(name="reactivate")
+@read_csv_arg(headers=_bulk_device_activation_headers)
+@format_option
+@sdk_options()
+def bulk_reactivate(state, csv_rows, format):
+    """Reactivate all devices from the provided CSV containing a 'guid' column."""
+    sdk = state.sdk
+    csv_rows[0]["reactivated"] = False
+    formatter = OutputFormatter(format, {key: key for key in csv_rows[0].keys()})
+
+    def handle_row(**row):
+        try:
+            _reactivate_device(sdk, row["guid"])
+            row["reactivated"] = "True"
+        except Exception as e:
+            row["reactivated"] = "False: {}".format(e)
+        return row
+
+    result_rows = run_bulk_process(
+        handle_row, csv_rows, progress_label="Reactivating devices:"
     )
     formatter.echo_formatted_list(result_rows)