+++ tests/integration/test_alerts.py
@@ -2,7 +2,6 @@ from datetime import datetime
 from datetime import timedelta
 
 import pytest
-from tests.integration import run_command
 
 
 begin_date = datetime.utcnow() - timedelta(days=20)
@@ -10,7 +9,9 @@ end_date = datetime.utcnow() - timedelta(days=10)
 begin_date_str = begin_date.strftime("%Y-%m-%d")
 end_date_str = end_date.strftime("%Y-%m-%d")
 
-ALERT_COMMAND = "code42 alerts search -b {} -e {}".format(begin_date_str, end_date_str)
+ALERT_SEARCH_COMMAND = "code42 alerts search -b {} -e {}".format(
+    begin_date_str, end_date_str
+)
 ADVANCED_QUERY = """{"groupClause":"AND", "groups":[{"filterClause":"AND",
 "filters":[{"operator":"ON_OR_AFTER", "term":"eventTimestamp", "value":"2020-09-13T00:00:00.000Z"},
 {"operator":"ON_OR_BEFORE", "term":"eventTimestamp", "value":"2020-12-07T13:20:15.195Z"}]}],
@@ -25,25 +26,27 @@ ALERT_ADVANCED_QUERY_COMMAND = "code42 alerts search --advanced-query '{}'".form
 @pytest.mark.parametrize(
     "command",
     [
-        ALERT_COMMAND,
-        "{} --state OPEN".format(ALERT_COMMAND),
-        "{} --state RESOLVED".format(ALERT_COMMAND),
-        "{} --actor user@code42.com".format(ALERT_COMMAND),
-        "{} --rule-name 'File Upload Alert'".format(ALERT_COMMAND),
-        "{} --rule-id 962a6a1c-54f6-4477-90bd-a08cc74cbf71".format(ALERT_COMMAND),
-        "{} --rule-type FedEndpointExfiltration".format(ALERT_COMMAND),
-        "{} --description 'Alert on any file upload'".format(ALERT_COMMAND),
-        "{} --exclude-rule-type 'FedEndpointExfiltration'".format(ALERT_COMMAND),
-        "{} --exclude-rule-id '962a6a1c-54f6-4477-90bd-a08cc74cbf71'".format(
-            ALERT_COMMAND
+        ALERT_SEARCH_COMMAND,
+        "{} --state OPEN".format(ALERT_SEARCH_COMMAND),
+        "{} --state RESOLVED".format(ALERT_SEARCH_COMMAND),
+        "{} --actor user@code42.com".format(ALERT_SEARCH_COMMAND),
+        "{} --rule-name 'File Upload Alert'".format(ALERT_SEARCH_COMMAND),
+        "{} --rule-id 962a6a1c-54f6-4477-90bd-a08cc74cbf71".format(
+            ALERT_SEARCH_COMMAND
         ),
-        "{} --exclude-rule-name 'File Upload Alert'".format(ALERT_COMMAND),
-        "{} --exclude-actor-contains 'user@code42.com'".format(ALERT_COMMAND),
-        "{} --exclude-actor 'user@code42.com'".format(ALERT_COMMAND),
-        "{} --actor-contains 'user@code42.com'".format(ALERT_COMMAND),
+        "{} --rule-type FedEndpointExfiltration".format(ALERT_SEARCH_COMMAND),
+        "{} --description 'Alert on any file upload'".format(ALERT_SEARCH_COMMAND),
+        "{} --exclude-rule-type 'FedEndpointExfiltration'".format(ALERT_SEARCH_COMMAND),
+        "{} --exclude-rule-id '962a6a1c-54f6-4477-90bd-a08cc74cbf71'".format(
+            ALERT_SEARCH_COMMAND
+        ),
+        "{} --exclude-rule-name 'File Upload Alert'".format(ALERT_SEARCH_COMMAND),
+        "{} --exclude-actor-contains 'user@code42.com'".format(ALERT_SEARCH_COMMAND),
+        "{} --exclude-actor 'user@code42.com'".format(ALERT_SEARCH_COMMAND),
+        "{} --actor-contains 'user@code42.com'".format(ALERT_SEARCH_COMMAND),
         ALERT_ADVANCED_QUERY_COMMAND,
     ],
 )
-def test_alert_command_returns_success_return_code(command):
-    return_code, response = run_command(command)
+def test_alert_command_returns_success_return_code(command, command_runner):
+    return_code, response = command_runner(command)
     assert return_code == 0